refactor folder structure for redux (application / component)
