+++ src/actionTypes/applicationActionTypes.ts
@@ -1,3 +1,4 @@
+// ** Application specific action types
 export const LOAD_INIT_DATA: string = 'LOAD_INIT_DATA';
 export const ADD_COMPONENT: string = 'ADD_COMPONENT';
 export const ADD_CHILD: string = 'ADD_CHILD';
@@ -7,9 +8,6 @@ export const DELETE_COMPONENT: string = 'DELETE_COMPONENT';
 export const CHANGE_FOCUS_COMPONENT: string = 'CHANGE_FOCUS_COMPONENT';
 export const CHANGE_COMPONENT_FOCUS_CHILD: string = 'CHANGE_COMPONENT_FOCUS_CHILD';
 export const CHANGE_FOCUS_CHILD: string = 'CHANGE_FOCUS_CHILD';
-export const UPDATE_CHILDREN: string = 'UPDATE_CHILDREN';
-export const REASSIGN_PARENT: string = 'REASSIGN_PARENT';
-export const SET_SELECTABLE_PARENTS: string = 'SET_SELECTABLE_PARENTS';
 export const EXPORT_FILES: string = 'EXPORT_FILES';
 export const EXPORT_FILES_SUCCESS: string = 'EXPORT_FILES_SUCCESS';
 export const EXPORT_FILES_ERROR: string = 'EXPORT_FILES_ERROR';
@@ -18,11 +16,6 @@ export const HANDLE_TRANSFORM: string = 'HANDLE_TRANSFORM';
 export const CREATE_APPLICATION: string = 'CREATE_APPLICATION';
 export const CREATE_APPLICATION_SUCCESS: string = 'CREATE_APPLICATION_SUCCESS';
 export const CREATE_APPLICATION_ERROR: string = 'CREATE_APPLICATION_ERROR';
-export const MOVE_TO_BOTTOM: string = 'MOVE_TO_BOTTOM';
-export const MOVE_TO_TOP: string = 'MOVE_TO_TOP';
-export const OPEN_EXPANSION_PANEL: string = 'OPEN_EXPANSION_PANEL';
-export const DELETE_PROP: string = 'DELETE_PROP';
-export const ADD_PROP: string = 'ADD_PROP';
 export const DELETE_ALL_DATA: string = 'DELETE_ALL_DATA';
 export const CHANGE_IMAGE_PATH: string = 'CHANGE_IMAGE_PATH';
 export const UPDATE_HTML_ATTR: string = 'UPDATE_HTML_ATTR';
