--- src/actionTypes/index.js
@@ -1,29 +1,0 @@
-export const LOAD_INIT_DATA = 'LOAD_INIT_DATA';
-export const ADD_COMPONENT = 'ADD_COMPONENT';
-export const ADD_CHILD = 'ADD_CHILD';
-export const DELETE_CHILD = 'DELETE_CHILD';
-export const UPDATE_COMPONENT = 'UPDATE_COMPONENT';
-export const DELETE_COMPONENT = 'DELETE_COMPONENT';
-export const CHANGE_FOCUS_COMPONENT = 'CHANGE_FOCUS_COMPONENT';
-export const CHANGE_COMPONENT_FOCUS_CHILD = 'CHANGE_COMPONENT_FOCUS_CHILD';
-export const CHANGE_FOCUS_CHILD = 'CHANGE_FOCUS_CHILD';
-export const UPDATE_CHILDREN = 'UPDATE_CHILDREN';
-export const REASSIGN_PARENT = 'REASSIGN_PARENT';
-export const SET_SELECTABLE_PARENTS = 'SET_SELECTABLE_PARENTS';
-export const EXPORT_FILES = 'EXPORT_FILES';
-export const EXPORT_FILES_SUCCESS = 'EXPORT_FILES_SUCCESS';
-export const EXPORT_FILES_ERROR = 'EXPORT_FILES_ERROR';
-export const HANDLE_CLOSE = 'HANDLE_CLOSE';
-export const HANDLE_TRANSFORM = 'HANDLE_TRANSFORM';
-export const CREATE_APPLICATION = 'CREATE_APPLICATION';
-export const CREATE_APPLICATION_SUCCESS = 'CREATE_APPLICATION_SUCCESS';
-export const CREATE_APPLICATION_ERROR = 'CREATE_APPLICATION_ERROR';
-export const MOVE_TO_BOTTOM = 'MOVE_TO_BOTTOM';
-export const MOVE_TO_TOP = 'MOVE_TO_TOP';
-export const OPEN_EXPANSION_PANEL = 'OPEN_EXPANSION_PANEL';
-export const DELETE_PROP = 'DELETE_PROP';
-export const ADD_PROP = 'ADD_PROP';
-export const DELETE_ALL_DATA = 'DELETE_ALL_DATA';
-export const CHANGE_IMAGE_PATH = 'CHANGE_IMAGE_PATH';
-export const UPDATE_HTML_ATTR = 'UPDATE_HTML_ATTR';
-export const UPDATE_CHILDREN_SORT = 'UPDATE_CHILDREN_SORT'
