--- src/reducers/index.js
@@ -1,9 +1,0 @@
-import { combineReducers } from 'redux';
-
-import componentReducer from './componentReducer.ts';
-
-const reducers = combineReducers({
-  workspace: componentReducer,
-});
-
-export default reducers;
