+++ src/reducers/index.ts
@@ -1,0 +1,11 @@
+import { combineReducers } from 'redux';
+
+import applicationReducer from './applicationReducer';
+import componentReducer from './componentReducer';
+
+const reducers = combineReducers({
+  workspace: applicationReducer,
+  component: componentReducer,
+});
+
+export default reducers;
